requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mutagen==1.47.0
pycparser==2.21
pycryptodomex==3.19.0
PyNaCl==1.5.0
python-dotenv==1.0.0
pytube==15.0.0
pytube3==9.6.4
typing_extensions==4.8.0
websockets==12.0
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2023.10.13
